Minor adjustment in XNES test
Lowering the number of required generations (with given seed)
after correctly the initialization value for {log_,}sigma.

Minor and not necessary, as the algorithm is scale invariant
anyway, but I'd rather keep the controls tight. This way the
initialization values are unitary and easier to scale later,
directly or indirectly (e.g. lrate).